[#3558] UnusedArgCorrector extraction
