+++ lib/rubocop/cop/lint/unused_method_argument.rb
@@ -23,6 +23,12 @@ module RuboCop
       class UnusedMethodArgument < Cop
         include UnusedArgument
 
+        def autocorrect(node)
+          UnusedArgCorrector.correct(processed_source, node)
+        end
+
+        private
+
         def check_argument(variable)
           return unless variable.method_argument?
           return if variable.keyword_argument? &&
